refacto&bugfix: passage des deux composants de Row en un seul, et débug du type de l'ID en Int
