+++ api/src/team/team.entity.ts
@@ -1,11 +1,11 @@
 import "reflect-metadata";
 import { Entity, Column, PrimaryGeneratedColumn, BaseEntity } from "typeorm";
-import { Field, ObjectType, ID } from "type-graphql";
+import { Field, ObjectType, Int } from "type-graphql";
 
 @ObjectType()
 @Entity()
 export class Team extends BaseEntity {
-  @Field(() => ID)
+  @Field(() => Int)
   @PrimaryGeneratedColumn()
   id: number;
